Brain Tumor Detection
The brain is the central part of Human body. Our emotions, memories, speech, ideas, and body action are all coordinated and managed by it . As a result, it's a significant part that necessitates utmost attention and accurate diagnosing whenever required. Brain disease is majorly due to irregular brain tissue growth that directly injure the brain shape which leads to unusual formation and cause brain cancer (tumor). An intracranial tumor is considered one of the most abnormal tissue growth and is one among most dangerous diseases in all stages of human life. Brain and Spinal cord are majorly affected by these tumors. 
Because it is such an important element of the human body, early tumor detection is essential. So, in this paper, We propose a three-stage pre-processing model that improves the quality of Magnetic Resonance Imaging Images and training is done with the data set using the new DCNN model for optimum diagnosis of Pituitary, Glioma, Meningioma and many more. We employed a normalization technique called Batch in this model to reduce the training and execution time while increasing the learning rate. significant results have been observed when the suggested model is compared to existing models.